+++ src/constants.js
@@ -423,16 +423,15 @@ export const REQUEST_ERROR_MESSAGE_TRANSLATION_KEYS = {
 export const DCB_INSTANCE_ID = '9d1b77e4-f02e-4b7f-b296-3f2042ddac54';
 export const DCB_HOLDINGS_RECORD_ID = '10cd3a5a-d36f-4c7a-bc4f-e1ae3cf820c9';
 
-export const SETTINGS_SCOPES = {
-  CIRCULATION: 'circulation',
-};
-
 export const TAGS_SCOPES = {
   TAG: 'ui-tags.tags.manage',
 };
 
 export const SETTINGS_KEYS = {
   GENERAL_TLR: 'generalTlr',
+  PRINT_HOLD_REQUESTS: 'PRINT_HOLD_REQUESTS',
+  PRINT_EVENT_LOG_FEATURE: 'printEventLogFeature',
+  TLR: 'TLR',
 };
 
 export const LOCATIONS_API = 'locations';