+++ src/routes/RequestsRoute.test.js
@@ -37,7 +37,6 @@ import {
 import RequestsRoute, {
   buildHoldRecords,
   getListFormatter,
-  getPrintHoldRequestsEnabled,
   getLastPrintedDetails,
   getFilteredColumnHeadersMap,
   urls,
@@ -497,7 +496,9 @@ describe('RequestsRoute', () => {
       },
       printHoldRequests: {
         records: [{
-          value: '{"printHoldRequestsEnabled": true}',
+          value: {
+            printHoldRequestsEnabled: true,
+          },
         }],
       },
       circulationSettings: {
@@ -1159,11 +1160,11 @@ describe('RequestsRoute', () => {
           ...defaultProps.resources,
           configs: {
             hasLoaded: true,
-            records: [
-              {
-                value: '{"createTitleLevelRequestsByDefault": true}',
-              }
-            ],
+            records: [{
+              value: {
+                createTitleLevelRequestsByDefault: true,
+              },
+            }],
           },
         },
       };
@@ -1579,30 +1580,6 @@ describe('RequestsRoute', () => {
 
         expect(NoValue).toHaveBeenCalled();
       });
-    });
-  });
-
-  describe('getPrintHoldRequestsEnabled', () => {
-    it('should return true when printHoldRequestsEnabled is true', () => {
-      expect(getPrintHoldRequestsEnabled({
-        records: [{
-          value: '{"printHoldRequestsEnabled": true}',
-        }],
-      })).toBeTruthy();
-    });
-
-    it('should return false when printHoldRequestsEnabled is false', () => {
-      expect(getPrintHoldRequestsEnabled({
-        records: [{
-          value: '{"printHoldRequestsEnabled": false}',
-        }],
-      })).toBeFalsy();
-    });
-
-    it('should return false when value absent', () => {
-      expect(getPrintHoldRequestsEnabled({
-        records: [],
-      })).toBeFalsy();
     });
   });