+++ src/deprecated/routes/RequestsRoute/RequestsRoute.js
@@ -82,6 +82,7 @@ import {
   getSelectedSlipDataMulti,
   selectedRowsNonPrintable,
   getNextSelectedRowsState,
+  generateRequestSearchCqlString,
 } from '../../../utils';
 import { getTlrSettings } from '../../utils';
 import packageInfo from '../../../../package';
@@ -318,7 +319,7 @@ class RequestsRoute extends React.Component {
         params: {
           query: makeQueryFunction(
             'cql.allRecords=1',
-            '(id=="%{query.query}" or requesterId=="%{query.query}" or requester.barcode=="%{query.query}*" or instance.title="%{query.query}*" or instanceId="%{query.query}*" or item.barcode=="%{query.query}*" or itemId=="%{query.query}" or itemIsbn="%{query.query}" or searchIndex.callNumberComponents.callNumber=="%{query.query}*" or fullCallNumberIndex=="%{query.query}*")',
+            generateRequestSearchCqlString(),
             {
               'title': 'instance.title',
               'instanceId': 'instanceId',
@@ -900,7 +901,7 @@ class RequestsRoute extends React.Component {
     const filterQuery = filters2cql(RequestsFiltersConfig, deparseFilters(this.getActiveFilters()));
 
     if (queryTerm) {
-      queryString = `(requesterId=="${queryTerm}" or requester.barcode=="${queryTerm}*" or item.title="${queryTerm}*" or item.barcode=="${queryTerm}*" or itemId=="${queryTerm}")`;
+      queryString = generateRequestSearchCqlString(queryTerm);
       queryClauses.push(queryString);
     }
     if (filterQuery) queryClauses.push(filterQuery);